feat: ignore processing other directories
If a folder in the .projectspy configured directory is not listed in
lanes config, then the content changes will be ignored.

A possible use case is archiving tickets. Anything you no longer want to
show in the UI can be placed in a folder such as `archive`. That way, it
will still be in the project, and git, but not show on the ui