refactor: migrate to WebC for UI
A 1:1 implementation of our Nunjucks based layout done in WebC. This
results in less plugins and support tooling, infavour of having WebC
handle it for us. The CSS and JS is parsed and bundled from each
component and then attached to the page. This allows for the removal of
fastify/static and fastify/view.

WebC is a slower approach as well, I'm noticing responses taking upwards
of 70ms. Testing with `performance.now()` around the intialization and
compile steps, which was around 30-50ms each request.

Nunjucks is fine and all, and likely a better choice in the long run.
But I wanted to work with WebC to get a feel for the approach.

Beware the documentation is incorrect regarding how to access data at
certain levels, but you can find the notes in the release log for
0.11.X. `$data.` as the prefix instead of this.