feat: add auto commit support
A larger change was necessary, Git / MacOS / Node case insentivity
mismatches made this a little tedious. Having to rename existing folders
and ensure things were registered with Git correctly. Otherwise, when
adding a file, it was fine, but updating a file was based on case
sentitive path in git, which mismatched node and macos. Fun times.

project: task in-progress/auto-commit-on-change.md updated

project: task done/auto-commit-on-change.md updated